% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvSimulations-package.R
\docType{package}
\name{EvSimulations-package}
\alias{EvSimulations}
\alias{EvSimulations-package}
\title{EvSimulations: Functions for Simulations for SiMPore Inc. Simulation Study on Protein Detection Rates in EVs}
\description{
This package contains functions that may be used to run simulations for the SiMPore Inc. simulation study on rates of detection of proteins in EVs. Description of how to run each individual simulation are contained in the README.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeDuttweiler/EvSimulations}
  \item Report bugs at \url{https://github.com/LukeDuttweiler/EvSimulations/issues}
}

}
\author{
\strong{Maintainer}: Luke Duttweiler \email{lduttweiler@hsph.harvard.edu}

}
\keyword{internal}
