% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationHelpers.R
\name{genFOVs}
\alias{genFOVs}
\title{Generate Field of Vision (FOV) Data}
\usage{
genFOVs(
  evCount = 1e+05,
  fovDist = rep(1, 31)/31,
  prevRates = rep(0.5, 10),
  bindRates = rep(1, 10),
  nonSpecBindRate = 0.005,
  maxProteins = 10,
  maxBinds = 10
)
}
\arguments{
\item{evCount}{A numeric value representing the total number of extracellular vesicles (EVs) to be simulated across all FOVs. Default is \code{1e5}.}

\item{fovDist}{A numeric vector representing the distribution of EVs across the FOVs. The values should sum to 1. The number of FOVs desired is length(fovDist). Default is \code{rep(1, 31)/31}.}

\item{prevRates}{A numeric vector representing the prevalence rates of proteins. Number of different proteins is length(prevRates) Default is \code{rep(.5, 10)}.}

\item{bindRates}{A numeric vector representing the binding rates of antibodies. Number of antibodies is length(bindRates). length(bindRates) and length(prevRates) must be equal. Default is \code{rep(1, 10)}.}

\item{nonSpecBindRate}{A numeric value representing the non-specific binding rate of antibodies. Default is \code{.005}.}

\item{maxProteins}{A numeric value specifying the maximum number of proteins that can exist in a single EV. Default is \code{15}.}

\item{maxBinds}{A numeric value specifying the maximum number of bindings allowed in a single EV. Default is \code{15}.}
}
\value{
A list of binary matrices, each corresponding to a FOV. In each matrix, rows represent EVs and columns represent the binding of antibodies.
}
\description{
This function generates information available for each of the 31 Fields of Vision (FOVs)
by simulating extracellular vesicles (EVs) with certain protein binding profiles. The function
utilizes parameters such as the number of EVs, prevalence rates, binding rates, and others to
generate a list of binary matrices indicating the binding profiles for each FOV.
}
\examples{
fovs <- genFOVs(evCount = 1e5, fovDist = rep(1, 31)/31, prevRates = rep(.5, 10),
               bindRates = rep(1, 10), nonSpecBindRate = .005, maxProteins = 10, maxBinds = 10)

}
