% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationHelpers.R
\name{rowShrink}
\alias{rowShrink}
\title{Row Shrink Function}
\usage{
rowShrink(binVec, maxSum)
}
\arguments{
\item{binVec}{A binary vector (numeric vector with values 0 and 1).}

\item{maxSum}{A numeric value specifying the upper bound for the sum of the elements in \code{binVec}.}
}
\value{
A binary vector where the number of \code{1}s is reduced to meet the specified maximum sum, if necessary.
If the sum of \code{binVec} is already less than or equal to \code{maxSum}, the original vector is returned.
}
\description{
This function takes a binary vector as input and reduces the number of \code{1}s in the vector at random positions
such that the sum of the vector does not exceed a specified maximum sum.
}
