% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario1.R
\name{scenario1}
\alias{scenario1}
\title{Run a Single Instance of SiMPore Scenario 1}
\usage{
scenario1(
  evCount = 1e+05,
  fovDist = rep(1, 31)/31,
  prevRates = rep(0.5, 10),
  bindRates = rep(1, length(prevRates)),
  nonSpecBindRate = 0.005,
  maxProteins = 10,
  maxBinds = 10,
  bloodDrawEV = 1e+08,
  typeIRate = 0.01,
  assumedNonSpecBindRate = nonSpecBindRate
)
}
\arguments{
\item{evCount}{A numeric value representing the total number of extracellular vesicles (EVs) to be simulated across all FOVs. Default is \code{1e5}.}

\item{fovDist}{A numeric vector representing the distribution of EVs across the FOVs. The values should sum to 1. The number of FOVs desired is length(fovDist). Default is \code{rep(1, 31)/31}.}

\item{prevRates}{A numeric vector representing the prevalence rates of proteins. Number of different proteins is length(prevRates) Default is \code{rep(.5, 10)}.}

\item{bindRates}{A numeric vector representing the binding rates of antibodies. Number of antibodies is length(bindRates). length(bindRates) and length(prevRates) must be equal. Default is \code{rep(1, 10)}.}

\item{nonSpecBindRate}{A numeric value representing the non-specific binding rate of antibodies. Default is \code{.005}.}

\item{maxProteins}{A numeric value specifying the maximum number of proteins that can exist in a single EV. Default is \code{15}.}

\item{maxBinds}{A numeric value specifying the maximum number of bindings allowed in a single EV. Default is \code{15}}

\item{bloodDrawEV}{A numeric value representing the total number of EVs in the blood draw, from which evCount is a small portion that is experimented on. Defaults to 1e8. Really shouldn't try to set higher than this.}

\item{typeIRate}{Numeric value indicating the family-wise type one error rate desired. Defaults to .01.}

\item{assumedNonSpecBindRate}{Numeric value indicating the assumed non-specific binding rate used in the hypothesis test. Defaults to \code{nonSpecBindRate}.}
}
\value{
A list with two components:
\describe{
\item{Results}{A data frame containing the true prevalence, observed prevalence, protein existence flag, test result flag, type 1 error flag, and type 2 error flag for each protein.}
\item{Params}{A list of the parameters used for generating the simulation results.}
}
}
\description{
This function runs a simulation for SiMPore Scenario 1, which analyzes field of view (FOV) data, and tests for protein prevalence using a Bonferroni-adjusted alpha threshold for hypothesis testing.
}
\seealso{
\link{genFOVs}
}
